refactor: 도메인 구조를 기능 단위로 재정리 text
- learning 라우터를 content/practice/progress/records/stage3 기능 패키지 옆으로 이동
- Stage 3 service/schema를 learning/stage3 하위로 이동
- classroom controller/dto 구조를 teacher_router, student_view_router, schemas로 단순화
- instruction controller 구조를 teacher_router, student_router로 단순화
- RAG를 비즈니스 도메인에서 infrastructure/rag로 이동
- 공통 seed/fixture 데이터 로더와 PDF 원본을 common/data로 이동
- system 도메인을 developer 운영 API 도메인으로 리네임
- 빈 패키지 표시용 __init__.py 제거
- 이동된 모듈 import와 구조 테스트 갱신
- persona 기반 재설계 문서의 패키지 구조 설명 갱신
