--- app/domains/classroom/controller/__init__.py
@@ -1,1 +1,0 @@
-"""Classroom HTTP controllers."""
