--- app/domains/classroom/controller/dto/__init__.py
@@ -1,1 +1,0 @@
-"""Classroom controller DTOs."""
