--- app/domains/learning/content/__init__.py
@@ -1,1 +1,0 @@
-"""Learning content catalog domain."""
