--- app/domains/learning/content/repositories/__init__.py
@@ -1,1 +1,0 @@
-"""Content catalog repository implementations."""
