--- app/domains/learning/controller/__init__.py
@@ -1,1 +1,0 @@
-"""Learning HTTP controllers."""
