--- app/domains/learning/stages/__init__.py
@@ -1,1 +1,0 @@
-"""Stage-specific learning domain services."""
