+++ tests/classroom/test_classroom_loader.py
@@ -1,4 +1,4 @@
-from app.data.data_loader import classroom_loader
+from app.common.data.data_loader import classroom_loader
 
 
 class FakeMongoClient:
