+++ tests/classroom/test_teacher_controllers.py
@@ -5,8 +5,8 @@ import pytest
 from fastapi import HTTPException
 
 from app.domains.auth.models import User
-from app.domains.classroom.controller.teacher_classroom_router import list_class_students, list_my_classes
-from app.domains.classroom.controller.teacher_student_view_router import get_student_records
+from app.domains.classroom.teacher_router import list_class_students, list_my_classes
+from app.domains.classroom.student_view_router import get_student_records
 
 
 def _user(user_id: str = "teacher_1", role: str = "teacher") -> User:
