+++ tests/structure/test_domain_structure.py
@@ -5,20 +5,18 @@ def test_domain_packages_are_importable():
     modules = [
         "app.common.security",
         "app.domains",
-        "app.domains.system",
-        "app.domains.system.indexing_service",
-        "app.domains.system.router",
+        "app.domains.developer",
+        "app.domains.developer.indexing_service",
+        "app.domains.developer.router",
         "app.domains.auth",
         "app.domains.auth.models",
         "app.domains.auth.repository",
         "app.domains.auth.router",
         "app.domains.auth.schemas",
         "app.domains.auth.service",
-        "app.domains.classroom.controller",
-        "app.domains.classroom.controller.dto",
-        "app.domains.classroom.controller.dto.response",
-        "app.domains.classroom.controller.teacher_classroom_router",
-        "app.domains.classroom.controller.teacher_student_view_router",
+        "app.domains.classroom.schemas",
+        "app.domains.classroom.teacher_router",
+        "app.domains.classroom.student_view_router",
         "app.domains.classroom.dependencies",
         "app.domains.classroom.models",
         "app.domains.classroom.repositories.base",
@@ -28,8 +26,8 @@ def test_domain_packages_are_importable():
         "app.domains.chat.router",
         "app.domains.chat.schemas",
         "app.domains.chat.service",
-        "app.domains.instruction.controller.student_assignment_router",
-        "app.domains.instruction.controller.teacher_instruction_router",
+        "app.domains.instruction.student_router",
+        "app.domains.instruction.teacher_router",
         "app.domains.instruction.generation",
         "app.domains.instruction.models",
         "app.domains.instruction.repositories.base",
@@ -38,12 +36,11 @@ def test_domain_packages_are_importable():
         "app.domains.instruction.schemas",
         "app.domains.instruction.service",
         "app.domains.learning",
-        "app.domains.learning.controller",
-        "app.domains.learning.controller.content_catalog_router",
-        "app.domains.learning.controller.student_learning_router",
-        "app.domains.learning.controller.student_progress_router",
-        "app.domains.learning.controller.student_records_router",
-        "app.domains.learning.controller.student_stage3_router",
+        "app.domains.learning.content.router",
+        "app.domains.learning.practice.router",
+        "app.domains.learning.progress.router",
+        "app.domains.learning.records.router",
+        "app.domains.learning.stage3.router",
         "app.domains.learning.content",
         "app.domains.learning.content.models",
         "app.domains.learning.content.repositories.base",
@@ -57,16 +54,15 @@ def test_domain_packages_are_importable():
         "app.domains.learning.schemas",
         "app.domains.learning.service",
         "app.domains.learning.student_schemas",
-        "app.domains.learning.stages",
-        "app.domains.learning.stages.stage3_schemas",
-        "app.domains.learning.stages.stage3_service",
+        "app.domains.learning.stage3.schemas",
+        "app.domains.learning.stage3.service",
         "app.domains.agent",
         "app.domains.agent.router",
         "app.domains.agent.schemas",
-        "app.domains.rag",
-        "app.domains.rag.retriever",
-        "app.domains.rag.schemas",
-        "app.domains.rag.service",
+        "app.infrastructure.rag",
+        "app.infrastructure.rag.retriever",
+        "app.infrastructure.rag.schemas",
+        "app.infrastructure.rag.service",
         "app.infrastructure.db.mongo.indexes",
     ]
